+++ client/src/apiRequests/getTweet.ts
@@ -1,53 +1,18 @@
-import axios, { AxiosResponse } from 'axios'
+import serverRequest from './apiRequests';
 
-const serverURL = "http://localhost:4000";
+/*
+This module get the id tweet from the server
+*/
 
 function getTweet(id: string){
 
-  var res: AxiosResponse<any, any>;
+  const route = "/twitter/getTweet"
 
   var body = {
     id : id,
   };
 
-  async function getRequest() {
-
-    await axios
-
-      .post(serverURL+'/twitter/getTweet', body)
-
-      .then(function (response) {
-        console.log("Request Answer")
-        console.log(response);
-        res = response.data;
-      })
-
-      .catch(function (error: any) {
-        console.log(error);
-      });
-
-      if (res) {
-          return res;
-      } else {
-          throw new Error('Unsuccessful request');
-      }
-    }
-
-  (async () => {
-
-      try {
-          const response = await getRequest();
-          console.log("Client Request");
-          console.dir(response, {
-              depth: null
-          });
-          console.log("Client Request End");
-          return response;
-
-      } catch (e) {
-          console.log(e);
-      }
-  })();
+  return serverRequest(route,body);
 
 }
