+++ client/src/apiRequests/getUserTimeline.ts
@@ -1,53 +1,19 @@
-import axios, { AxiosResponse } from 'axios'
+import serverRequest from './apiRequests';
 
-const serverURL = "http://localhost:4000";
+/*
+This module get the id user's timeline
+*/
 
 function getUserTimeline(id: string, p_token: string = ""){
 
-  var res: AxiosResponse<any, any>;
+  const route = "/twitter/UserTimeline"
 
   var body = {
     id : id,
-    p_token : p_token,
+    p_token: p_token,
   };
 
-  async function getRequest() {
-
-    await axios
-
-      .post(serverURL+'/twitter/userTimeline', body)
-
-      .then(function (response) {
-        console.log("Request Answer")
-        console.log(response);
-        res = response.data;
-      })
-
-      .catch(function (error: any) {
-        console.log(error);
-      });
-
-      if (res) {
-          return res;
-      } else {
-          throw new Error('Unsuccessful request');
-      }
-    }
-
-  (async () => {
-
-      try {
-          const response = await getRequest();
-          console.log("Client Request");
-          console.dir(response, {
-              depth: null
-          });
-          console.log("Client Request End");
-
-      } catch (e) {
-          console.log(e);
-      }
-  })();
+  return serverRequest(route,body);
 
 }
